Moved some methods from App to Paint
